Restructure Flask
