+++ api/health.py
@@ -15,12 +15,11 @@
 #  You should have received a copy of the GNU Affero General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from .resource import ResourcesApi, ResourceApi
-from .root import Root, Health
+from flask import Response
+from flask_restful import Resource as Restful_Resource
 
 
-def initialize_routes(api):
-    api.add_resource(Root, '/')
-    api.add_resource(Health, '/health')
-    api.add_resource(ResourcesApi, '/resources')
-    api.add_resource(ResourceApi, '/resources/<resource_id>')
+class HealthApi(Restful_Resource):
+    def get(self):
+        # Health check
+        return Response(status=200)
